feat: route present logic moved to RoutePresenter
